Add one script per model
