+++ nlpsandboxclient/data_node_models/datasets.py
@@ -1,0 +1,8 @@
+"""Dataset"""
+
+
+class Dataset(Model):
+    """Dataset model"""
+    pass
+
+
