+++ nlpsandboxclient/data_node_models/fhir_stores.py
@@ -1,0 +1,9 @@
+"""FHIR store"""
+
+
+class FhirStore(Model):
+    """FhirStore model"""
+
+    @property
+    def dataset_id(self):
+        return self.kwargs.get("dataset_id")
